Add signalR and messageHub to make a simple communication with users in specific chat (group)

+++ real-time-online-chats.Api/real-time-online-chats.Server/Controllers/V1/ChatsController.cs
@@ -1,13 +1,17 @@
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
 using real_time_online_chats.Server.Contracts.V1;
 using real_time_online_chats.Server.Contracts.V1.Requests.Chat;
 using real_time_online_chats.Server.Contracts.V1.Responses;
+using real_time_online_chats.Server.Contracts.V1.Responses.Auth;
 using real_time_online_chats.Server.Contracts.V1.Responses.Chat;
 using real_time_online_chats.Server.Contracts.V1.Responses.Message;
 using real_time_online_chats.Server.Domain;
 using real_time_online_chats.Server.Extensions;
+using real_time_online_chats.Server.Hubs;
+using real_time_online_chats.Server.Hubs.Clients;
 using real_time_online_chats.Server.Services.Chat;
 
 namespace real_time_online_chats.Server.Controllers.V1;
@@ -16,32 +20,39 @@ namespace real_time_online_chats.Server.Controllers.V1;
 public class ChatsController : ControllerBase
 {
     private readonly IChatService _chatService;
+    private readonly IHubContext<MessageHub, IMessageClient> _messageHub;
 
-    public ChatsController(IChatService chatService)
+    public ChatsController(IChatService chatService, IHubContext<MessageHub, IMessageClient> messageHub)
     {
         _chatService = chatService;
+        _messageHub = messageHub;
     }
 
     [HttpGet(ApiRoutes.Chats.GetAll)]
-    public async Task<IActionResult> GetAll(int page = 1, int pageSize = 5)
+    public async Task<IActionResult> GetAll([FromQuery] int page = 1, [FromQuery] int pageSize = 5, [FromQuery] int countOfMessages = 3)
     {
         if (page <= 0 || pageSize <= 0)
         {
-            throw new ArgumentException("Page number and page size must be greater than zero.");
+            return BadRequest(new { Errors = "Page number and page size must be greater than 0." });
+        }
+
+        if (countOfMessages < 0 || countOfMessages > 4)
+        {
+            return BadRequest(new { Errors = "Count of messages must be greater than or equal to 0." });
         }
 
         var paginationChats = await _chatService.GetChatsAsync(page, pageSize);
-        var response = new PaginatedChatResponse
+        var response = new PaginatedResponse<GetChatResponse>
         {
             TotalCount = paginationChats.TotalCount,
             PageNumber = paginationChats.PageNumber,
             PageSize = paginationChats.PageSize,
-            Chats = paginationChats.Items.Select(c => new GetChatResponse
+            Items = paginationChats.Items.Select(c => new GetChatResponse
             {
                 Id = c.Id,
                 OwnerId = c.OwnerId,
                 Title = c.Title,
-            }).ToList(),
+            }),
         };
 
         return Ok(response);
@@ -52,7 +63,7 @@ public class ChatsController : ControllerBase
     {
         var chat = await _chatService.GetChatByIdAsync(chatId);
         if (chat is null) return NotFound();
-        
+
         var response = new GetChatResponse
         {
             Id = chat.Id,
@@ -63,26 +74,46 @@ public class ChatsController : ControllerBase
         return Ok(response);
     }
 
-    [HttpGet(ApiRoutes.Chats.GetRestrict)]
-    public async Task<IActionResult> GetRestrict([FromRoute] Guid chatId)
+    [HttpGet(ApiRoutes.Chats.GetDetailed)]
+    public async Task<IActionResult> GetDetailed([FromRoute] Guid chatId)
     {
         if (!HttpContext.TryGetUserId(out var userId)) return Unauthorized();
+        if (!await _chatService.IsUserExistInChatAsync(chatId, userId)) return Forbid();
 
-        var chat = await _chatService.GetChatRestrictByIdAsync(chatId, userId);
+        var chat = await _chatService.GetChatByIdWithDetailsAsync(chatId);
         if (chat is null) return NotFound();
-        
-        var response = new GetChatRestrictResponse
+
+        var response = new GetChatDetailedResponse
         {
             Id = chat.Id,
             OwnerId = chat.OwnerId,
             Title = chat.Title,
-            Members = chat.Members,
-            Messages = chat.Messages.Select(m => new GetMessageResponse
+            CreationTime = chat.CreationTime,
+            Messages = chat.Messages.Select(m => new GetMessageDetailedResponse
             {
                 Id = m.Id,
                 Content = m.Content,
-                UserId = m.UserId,
-            }).ToList(),
+                User = new UserResponse
+                {
+                    Id = m.UserId,
+                    Email = m.User.Email,
+                    FirstName = m.User.FirstName,
+                    LastName = m.User.LastName,
+                },
+            }),
+            Users = chat.Members.Select(m => new UserResponse
+            {
+                Id = m.Id,
+                Email = m.Email,
+                FirstName = m.FirstName,
+                LastName = m.LastName,
+            }).Append(new UserResponse
+            {
+                Id = chat.Owner.Id,
+                Email = chat.Owner.Email,
+                FirstName = chat.Owner.FirstName,
+                LastName = chat.Owner.LastName,
+            })
         };
 
         return Ok(response);
@@ -97,13 +128,15 @@ public class ChatsController : ControllerBase
         {
             OwnerId = userId,
             Title = request.Title,
+            Members = [.. request.UsersIds.Select(i => new UserEntity { Id = i })]
         };
 
         var created = await _chatService.CreateChatAsync(chat);
         if (!created) return BadRequest(new { Message = "Failed to create chat. Please try again." });
-        
+
         var response = new CreateChatResponse
         {
+            Id = chat.Id,
             OwnerId = chat.OwnerId,
             Title = chat.Title,
         };
@@ -114,6 +147,9 @@ public class ChatsController : ControllerBase
     [HttpPut(ApiRoutes.Chats.Update)]
     public async Task<IActionResult> Update([FromRoute] Guid chatId, [FromBody] UpdateChatRequest request)
     {
+        if (!HttpContext.TryGetUserId(out var userId)) return Unauthorized();
+        if (!await _chatService.IsUserOwnsChatAsync(chatId, userId)) return Forbid();
+
         var chat = new ChatEntity
         {
             Id = chatId,
@@ -134,7 +170,33 @@ public class ChatsController : ControllerBase
     [HttpDelete(ApiRoutes.Chats.Delete)]
     public async Task<IActionResult> Delete([FromRoute] Guid chatId)
     {
+        if (!HttpContext.TryGetUserId(out var userId)) return Unauthorized();
+        if (!await _chatService.IsUserOwnsChatAsync(chatId, userId)) return Forbid();
+
         var deleted = await _chatService.DeleteChatAsync(chatId);
         return deleted ? NoContent() : NotFound();
+    }
+
+    [HttpPost(ApiRoutes.Chats.Join)]
+    public async Task<IActionResult> Join([FromRoute] Guid chatId)
+    {
+        if (!HttpContext.TryGetUserId(out var userId)) return Unauthorized();
+        if (await _chatService.IsUserExistInChatAsync(chatId, userId)) return Ok();
+
+        var joined = await _chatService.UserJoinChatAsync(chatId, userId);
+        await _messageHub.Clients.Group(chatId.ToString()).JoinChat(userId);
+
+        return Ok(joined);
+    }
+
+    [HttpPost(ApiRoutes.Chats.Leave)]
+    public async Task<IActionResult> Leave([FromRoute] Guid chatId)
+    {
+        if (!HttpContext.TryGetUserId(out var userId)) return Unauthorized();
+
+        var leaved = await _chatService.UserLeaveChatAsync(chatId, userId);
+        await _messageHub.Clients.Group(chatId.ToString()).LeaveChat(userId);
+
+        return Ok(leaved);
     }
 }
